requirements.txt
ipykernel
pydantic 
python-dotenv
langgraph
awscli
boto3 
openai
langchain_nvidia_ai_endpoints
langchain_community
streamlit
faiss-cpu
langchain_core
langchain 
typing_extensions 
langchain-openai 
pyppeteer 
arxiv
wikipedia 
langchain-tavily